feat: do brower side checks in the inpage script itself do browser side checks + extension side cheks such as blocklist in the provider's content script signed-off-by: pavan joshi <pavanj914@gmail.com>
