+++ src/extension/inpage-script/shouldInject.js
@@ -1,0 +1,12 @@
+import {
+  doctypeCheck,
+  documentElementCheck,
+  suffixCheck,
+} from "../shouldInjectBrowserChecks";
+export default function shouldInjectInpage() {
+  const isHTML = doctypeCheck();
+  const noProhibitedType = suffixCheck();
+  const hasDocumentElement = documentElementCheck();
+
+  return isHTML && noProhibitedType && hasDocumentElement;
+}
